Alter ReplayInfo model bucket_path field to file_hash (Unique Identifier)
file_hash is now just the SHA256 hash of the uploaded file instead of the path to the bucket. Unlike the previous identifier (Battletag), it provides a unique identifier for replays. There is no need to store the path to the file in the bucket since all the information needed to construct the path is in the database.