Restructure views, alter URL paths and add redirecting for unauthenticated users
Battlenet_api app was removed and it's battlenet authentication functionality added to the user_profile app, since authentication is related to user profiles. This was done to clean up the structure of apps.

Multiple URL paths were changed from being includeded in site_structure to being in the project urls file due to increased functionality of the apps.

Users that have not authenticated their battlenet account are now redirected to an authentication sub page when trying to access the profile page. Implemented using @user_passes_test decorator