Create new profile page template. Begin refactoring of CSS
Made a new prototype of the user profile page that fits the landing page theme. Is relatively stable, will dynamically resize with the page (To a certain point) and automatically centers headers over the content. Need to do user testing about the side nav since the area to activate it might be too small.