Add token auth and permissions to API and alter DB models
The API now requires users trying to access it to have been authenticated using Django's built in TokenAuthentication. Still need to implement throttling and other permissions.

The database models have been altered in multiple ways. The following is not an exhaustive list.

- Natural keys replaced with surrogate keys in multiple tables
- Renamed and altered existing replay model
- Added new replay model

Natural keys were replaced with surrogate as certain keys (I.e. Email) were deemed not appropriate to be a PK.
The replay model was altered and renamed to allow support for different gamemodes in future and to better reflect the data it stores.
The new replay model represents all replays uploaded that aren't tied to an authenticated battlenet account. They are instead simply tied to the user's main account.