Create prototype replay verification script
The verification script searches all current unauthenticated replays the user has and checks to see if they are related to an authenticated battlenet account. If a replay is related, copy all the data to the AuthenticatedReplay table and delete the UnauthenticatedReplay record. Need to add bucket object renaming and store more game data in replay tables for script to work, though.