Alter replay models to prevent possible PK collisions
Since there is a possibility that 2 players upload the same game and the fact that hashes are technically not collision proof, the 'file_hash' field is no longer the PK in the replay models.

Consequences of this change are: Changed logic when checking for duplicate replay to only detect a duplicate if it's from the same account and the database was reset since Django wasn't playing nice. Bye bye migrations.