Improve First-time Login Flow + Add Profiles Manually (#3)
* Refactor /login and /authorize/check endpoints to return all user info

* Add /profile endpoint for parsing and saving profiles

The /profile endpoint recieves a URL from the frontend that is parsed with regex to extract the fragment containing profile info. That fragment is then inserted into an API URL, where the full profile info in JSON format is fetched. All the relevant profile info is then consolidated in a dictionary and saved to the users BattlenetAccount model. If data for the profile already exists the dictionaries are updated or merged.