Create FeatureVote model and /vote/ endpoint
The FeatureVote model contains information about an individual vote from any set of options. It is tied to a set of options through vote_id and contains vote information in the feature and comment fields. The time at which the vote was submitted is also recorded. Attempting to submit more than 2 votes causes the extra votes to be dropped, any invalid feature code also causes votes to be dropped. Votes are managed such that there only exists 2 votes per user in the database after the request is complete.