Create replays/summary/ endpoint + modify filter_user_replays
The replays/summary/ endpoint returns information on the number of linked and unlinked replays attached to a user account. The filter_user_replays function was modified to be more flexible and handle this summary endpoint as well as the existing count endpoint. The replays/all/ endpoint was removed due to it being dead code.